Fixed issue with horizontal and vertical checking functions
Before, the horizontal and vertical checking functions did not check the middle row/column (respectively) if it exists in a selection with an odd number of rows/columns (respectively).
This may not have ever been an issue in a real use case, but it is better to fix it now.
This could only be an issue if the Glyph at this location has a non-self-symmetric connection type. I haven't been able to generate a situation where this happens, but it's better to be safe.