Finished generating both diagonal symmetries and full symmetry
Abandoned the idea of generating symmetries of "diagonal plus extra" since that's too much work in a `tryGenerating` function. Might be added in future.

Right now, `Knot::generateFullSym()` actually works on some selections that are off-from-square by 1 unit in a direction. This will be changed to work for all non-square selections.