Fixed `Knot::checkHoriSym()` and `Knot::checkVertSym()` to use inside connections instead of outside connections
This makes it work for the wrapping case as well.

That being said, this uncovered another problem. If a wrapped knot has been generated, then a non-wrapped knot can be generated anywhere, which may break the wrapping conditions. To do, write a new function `Knot::checkWrapping()` to take care of this issue.